requirements.txt
streamlit
openai
pytesseract
pdfplumber
pillow
python-dotenv
